Adding routes directory and moving routers to routes folder
